third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGMaskElement-svgdom-y-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'y' property of the SVGMaskElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS maskElement.y.baseVal.valueAsString is "-100%"
PASS maskElement.y.baseVal.valueAsString is "-10%"
PASS successfullyParsed is true

TEST COMPLETE
